[lcd4linux @ 2006-09-04 16:41:21 by siretart] removing debianisation from upstream source, as discussed with Michael on private mail.
Rationale: I try to keep an updated lcd4linux in debian/sid. If you need the debinisation,
please use my packages as base. If you want more updated package, please drop me a note,
and I'll arrange another upload to unstable (or experimental, depending on the state
of the cvs).

git-svn-id: https://ssl.bulix.org/svn/lcd4linux/trunk@704 3ae390bd-cb1e-0410-b409-cd5a39f66f1f